ref(autoscaling): expose different metric name (#4618)
This PR exposes the same values as `relay_utilization` using the key
`relay_service_utilization`.

With the addition of new `relay_..._utilization` metrics, the service
related metric `relay_utilization` is not very descriptive and can lead
to confusion.
To enable a smooth transition away from the generic name, we will
1. expose the new metric name
2. change everything that relies on the old name to the new name
3. remove the exposure of the old name