feat(span): serialize span _meta over kafka
Events have a `_meta` property that contains information on transformations
made in Relay to incoming data. Sentry is deprecating events and using EAP as
the source of truth for all span information, so spans need their related
`_meta` information too for the FE to eventually use. The first step is to
write the info to Kafka so that it can be written to EAP downstream.