ref(test): don't rely on processing for DSC propagation test (#4645)
Responding to @Dav1dde's comment:
https://github.com/getsentry/relay/pull/4630#discussion_r2027293915

> If you make this into a 'pop' test with `relay(mini_sentry)` the test
will run significantly faster (and less flaky). To read the data you
then can use `mini_sentry.captured_events`.

The old test relied on the materialized `_dsc` property on the event,
but that is only materialized when processing is enabled so the test now
checks the envelope header instead. This means the test covers a little
less code, but trades it off for speed and reliability.