feat: Make default Apple device class high (#4609)
Change the default device class to high for iPhones and iPads. We have
an extensive list of these devices, which are well-known. If a device is
not on this list, it must be a new one. Previously, we set the device
class to unknown until we manually update this list. Now, the default is
high.

This came up while investigating
https://github.com/getsentry/sentry/issues/68258.