fix(otlp): add OTLP trace endpoint without trailing slash (#4655)
We structured our OTLP endpoint URL to match the OTLP standard, but we
added a trailing slash that [isn't present in the
spec](https://opentelemetry.io/docs/specs/otlp/#otlphttp-request). That
causes OTLP clients that automatically apply the `/v1/traces` suffix to
404 when trying to reach Relay. Add a version without the trailing slash
to match the spec.

Out of an abundance of caution this PR keeps the version with the
trailing slash just in case someone is using it (although there are no
known clients).