feat(relay): add settings to infer ip (#4623)
This PR adds a `settings` structure in the SDK context of an event and
adds the `infer_ip` field to give explicit control if an IP address
should be inferred.

Currently 3 values are supported, one of them being used to describe
legacy behavior
* `auto` - equivalent to the old `{{auto}}`, it will infer the IP
address from request information unless `user.ip_address` contains an
actual IP address. The main difference is that it will also infer the IP
if the submitted IP address is missing.
* `never` - Do no derive the IP under any circumstances. If
`user.ip_address` contains `{{auto}}`, it will be removed.
* `legacy` - If no value is sent, then this will be the default. Uses
old behaviour of inferring the IP if `{{auto}}` is specified and also
backfills `REMOTE_ADDR`. It will also enable the IP inference for
`javascript`, `cocoa` and `objc` if the submitted IP address is missing