ci(deploy): Also check that artifact upload has completed (#4671)
As part of the deploy pipeline we call this
[script](https://github.com/getsentry/devinfra-deployment-service/blob/fa1e4bd252feb75b18b7b0d703da7b601549ca05/gocd_agent/scripts/checks/githubactions/checkruns.py#L126)
which ensures that the 'github runs' that are passed (as arguments to
the script) are completed before the script terminates.

Previously we did not wait for the upload of the artifacts to GOCD which
could cause the
[`create-sentry-release`](https://github.com/getsentry/relay/blob/master/scripts/create-sentry-release)
script to fails as the files it tries for pull from GOCD were not yet in
GOCD. (This only happens though if you triggered the pipeline very
shortly after the CI 'finished').

This PR adds the logic to also wait for the upload of the artifacts to
GOCD such that it is impossible for the
[`create-sentry-release`](https://github.com/getsentry/relay/blob/master/scripts/create-sentry-release)
script to run (and fail) before the upload of the artifacts has
finished.

#skip-changelog