feat(links): write span links to Kafka (#4601)
Span links were added in #4486, but are currently only accessible as
part of transaction events. We would like to get them in EAP. The first
step in that process is to add them to serialized Kafka spans.

The integration tests confirm production into Kafka for every current
path for links to enter the system:
- in a transaction event's trace context
- in a transaction event's child spans
- in a `span` envelope item
- in an `otel_span` envelope item
- via OTLP JSON
- via OTLP gRPC

Note that our version of the `opentelemetry-proto` Python library is too
old to make `flags` available so not all tests could check the `sampled`
handling. I'll handle that in a future PR.

---------

Co-authored-by: Lukas Stracke <lukas.stracke@sentry.io>
Co-authored-by: Sigrid Huemer <32902192+s1gr1d@users.noreply.github.com>