ref(project-cache): Add even more detailed fetch metrics (#4719)
Breaks down the latency by the delay, a latency with a high delay is not
sign of an issue, it's just a sign of a low volume project. But
depending on how often this happens, we may want to address this case
too.

Also tracks 'end to end' duration of a fetch, through the entire system,
from initiating the fetch to completing it. Long fetch are always a sign
for a problem in the system, but they are mostly dominated by upstreams
not being able to provide the project config.

Refs: INGEST-180