feat(config): Allow environment references in Relay configuration (#4750)
> Allow ${VAR} references in Relay's config. Any leaf value can now be a
reference to an environment variable.

Another attempt at #4748 for #4647, now with version 0.2 (instead of
0.1).

serde-vars 0.2 is now more lenient in deserialization when it can (for
strings and bytes) and the underlying deserializer supports it. This
will fix the config issue with YAML configs, which broke self hosted
CI/the self hosted default Kafka config.