feat: Implement conversion from OTEL to span V2 (#4752)
This implements the conversion from OTEL spans to spans V2 and replaces
the implementation of `otel_to_sentry::otel_to_sentry_span` by the
composition of `OTEL -> v2` and `v2 -> v1`.

In the course of this, it also moves some conversion logic from `v2 ->
v1` to `OTEL -> v2`. In particular, the `v2 -> v1` conversion now
assumes that the `name` and `description` of the V2 span are already
normalized and correct. I've documented which fields are set based on
which attributes on each of the conversion functions.