ref: Represent SpanId as bytes (#4756)
This changes the representation of `SpanId` from `String` to `[u8; 8]`.
The constructor is now private; the only ways to construct a `SpanId`
are the `FromStr` and `TryFrom<&[u8]>` impls, both of which check
that the ID has the right length and is not nil.

In order to handle `SpanId`s in `Getter` impls, it adds a new `HexId` struct
for handling hexadecimal IDs represented as bytes. The variant `Val::Uuid`
is replaced with `Val::HexId` that handles such `HexId`s.