ref(relay): Split envelopes only in processing instead of all over (#4762)
Instead of splitting the envelope in the endpoint *and* buffer and
afterwards relying on properly set envelope groups in the processor, the
splitting is now moved into the processor.

This is the first refactor to allow for a more specialized and typed
envelope processor, which is supposed to take items out of an envelope
on demand to create a typed 'unit of work'. Such a design is in the
current Relay design not possible, as buffering requires an envelope.

As a side effect, with this change we actually also optimize the
envelope buffer a bit, since we store larger envelopes, but fewer.

Most changes are due to the processing group no longer being directly
tied to the envelope. This makes sense as the envelope never guaranteed
what items it contained. It is now kept outside the envelope and fused
into the envelope when constructing the `TypedEnvelope`.