ref(spans): Remove SpanKind::Unspecified (#4774)
Since the `kind` field of `Span` is not mandatory, and the `Unspecified`
variant of `SpanKind` transports no more information than an empty
`kind`, we remove `SpanKind::Unspecified` and use `Annotated::empty`
instead.

Moreover, we remove `SpanV2Kind::Other` because the list of kinds is
very unlikely to be extended in the future.

In detail this means:

* When converting from OTEL to Span V2:
  * Both `Unspecified` and an empty span kind map to `empty`
* All other valid OTEL span kinds map to the corresponding V2 span kinds
  * All other `i32` values map to errors
* When converting from Span V2 to Span V1:
  * All variants of `SpanV2Kind` map to the corresponding V1 kinds
* When converting from OTEL to Span V1:
  * Both `Unspecified` and an empty span kind map to `empty`
* All other valid OTEL span kinds map to the corresponding V1 span kinds
  * All other `i32` values map to errors