ref(kafka): Make routing key optional, instead of randomizing it  (#4765)
Another attempt at #4738, but this time it excludes the metrics topic.

As we've seen not setting an explicit message key reduces the load on
Kafka quite significantly, but due to an issue with the metrics indexer
in Sentry, we cannot enable it for this specific topic.

The change in general is desirable, from a Kafka PoV as well as Relay
PoV.