feat(ourlogs): Add browser name/version to logs (#4757)
We've received feedback externally and internally that users want to see
browser information attached to their frontend logs. This matches
behaviour that already exists in other data types like spans and errors
(events) in the product.

To accomplish this, we generate new attributes on the log based on the
incoming user agent and client hints. This matches the approach we take
with spans and errors.

We set these via two new attributes, `browser.name` and
`browser.version`. This is documented in the conventions here:
-
https://github.com/getsentry/sentry-conventions/blob/main/model/attributes/browser/browser__name.json
-
https://github.com/getsentry/sentry-conventions/blob/main/model/attributes/browser/browser__version.json

resolves https://linear.app/getsentry/issue/LOGS-133