fix(ourlog): use sentry prefix for browser name/version (#4783)
In https://github.com/getsentry/relay/pull/4757 we decided to attach
browser name and version to logs.

Right now for spans, we prefix browser information with `sentry.X`, see
https://github.com/getsentry/sentry/blob/34c738613c9b39e3b0d64336d13b4551a4075c66/src/sentry/search/eap/spans/attributes.py#L334.

This is causing inconsistencies between the queries for spans and logs,
which is breaking the current queries for the logs view. Considering we
are "enhancing" the data via relay, I think it makes sense to also add a
`sentry.X` prefix for logs. This PR makes that change.

resolves https://linear.app/getsentry/issue/LOGS-152