fix(otlp): Propagate all attributes (#4779)
In the process of converting OTLP spans to `SpanV2` and then `SpanV1`, a
few important attributes get dropped. This happens because during the
matching process to extract the attributes, they are not written to the
data bag. This PR updates the logic to make sure that any attributes (as
long as they're not top-level `Span` fields) are written.

#skip-changelog