fix(otlp): Attach span name as an attribute to `Span` messages (#4780)
The plan for `name`, `op`, and `description` is in [Span Field
Transition](https://linear.app/getsentry/document/span-field-transition-e0e4b8e59f11).
The tl;dr is that we're going to keep all those fields separate. Right
now, the `span_v2_to_span_v1` puts the `name` of the incoming span into
the `op` of the outgoing span, which is confusing.

Instead, we're going to:

1. Derive a proper `op` and `description` for the incoming V2 span based
on its attributes, and write that into the `op` and `description` fields
of the outgoing V1 span, respectively. The buffer will use the alias map
to map those fields to `sentry.op` and `sentry.description` attributes
and send it to EAP. This part is coming in other PRs
2. Take the incoming `name` field of the V2 span, and write that to the
attribute `sentry.name`. The buffer will leave this attribute alone, and
pass it to EAP. That's what this PR is doing. It creates temporary
duplication since the `op` field of the outgoing V1 span and the
`sentry.name` attribute of the outgoing span have the same data, but
that will be resolved soon when we derive a proper `op` for it

#skip-changelog

---------

Co-authored-by: Sebastian Zivota <loewenheim@users.noreply.github.com>