feat(profiling): strip out profiler_id from profile context for short transactions (#4818)
For continuous profiles, we want to make sure that, in case the
transaction duration is `< 19.8 ms`, the `profiler_id` field in the
`profile context` is stripped out.

The reason for this is that in that case, we know we wouldn't have
enough "samples" (101hz freq.) and the profile data associated with it
would be unusable.