ref: Use FiniteF64 for measurements (#4828)
The `FiniteF64` type is intended for `f64` values that can't be ±∞ or
NaN. It's currently only used for metrics, but measurements ought to
work the same way. Therefore, this PR changes measurements to `FiniteF64`.

* It moves `FiniteF64` to `relay-protocol` so that all parts of Relay
that need it have access to it.
* It implements a bunch of traits for `FiniteF64`: `FromValue`,
`IntoValue`, `ProcessValue`, `Empty`, `PartialOrd<f64>`,
`PartialEq<f64>`, and `AddAssign`. The only notable thing here is that
in order to implement `ProcessValue` I added a new method
`process_finite_f64` to the `Processor` trait, as simply reusing
`process_f64` wouldn't be sound.

Since measurement values were `Annotated<f64>` before, it costs
us very little to change them to `Annotated<FiniteF64>` and use the
`Annotated` for invalid result values.