ref(server): Introduces a new processing pipeline and implements it for logs (#4777)
Introduces a new processing pipeline and implements it for logs.
All of the log processing logic has been moved to the new pipeline.

A `Processor` is designed to build its unit of work from multiple
sources, for now it only supports a `ManagedEnvelope` as this is what we
do today. In the future it's possible to imagine a custom transport
between Relays and processor will be able to process from different
formats.

Similar to the input, the output is now also an additional layer which
can be serialized into multiple formats, for now this only supports
`ManagedEnvelope` again, but in the future, an additional serialization
will be introduced to serialize to a format which the `StoreService`
(Kafka) expects. This finally means we no longer need to needlessly
`serialize->deserialize->serialize` certain items for Kafka.

Rate limiting is now no longer envelope based, which makes it much much
easier, but has the downside that envelope based rate limiting is still
currently necessary for the fast path. We might want to explore more in
this direction in the future.

The processor also forces a correct outcome handling, through the type
system (usage of `Rejected<>`), which is much easier to use. As well as
just runtime, debug validations of outcome counts. For every change in
items must be accounted for.

The processing code itself is structured to include more strongly typed
containers, instead of an untyped envelope with a lot of implicit state.

Processing is now longer conditionally compiled, ideally all of this
should be hidden behind either normal `if` statements or a facade which
conditionally does nothing or directs to somewhere else (like the new
rate limiter). This means we can maybe finally reduce the cancer which
is `cfg(feature = "processing")`, which is only really useful for Kafka
and Redis dependencies.

Refs: INGEST-395