feat(ai): Mark some AI attributes as PII maybe (#4860)
The default scrubbers are too eager on these fields, rendering the
experience often unusable.
We're moving the specified attributes to PII maybe which means that
we'll be not scrubbing them by default, but people will be able to add
it to scrubbing in settings if they wish to do so. This will be properly
documented in product docs.

---------

Co-authored-by: David Herberth <david.herberth@sentry.io>
Co-authored-by: Vjeran Grozdanic <vjeran.grozdanic@sentry.io>