feat(ai): Conditionally set `total_cost` and `total_tokens` attributes (#4868)
Currently we always set `total_cost` and `total_tokens` attributes on
every AI span, even if `input_tokens` or `output_tokens` are not
present.

At the moment, that is not the behavior we want, since for example we
have spans that are just invoking agent and it has no cost nor any
tokens are used, so setting this attributes is wrong.

This also shows up then in UI and additionally confusing users