ref(store): Reduce owned deserialization (#4845)
Some fields of `SpanKafkaMessage` were previously deserialized as
`serde_json::Value`, which was partly caused by having to parse and
validate `f64` values in `measurements`. As of #4828 that validation is
no longer necessary. Consequently, we can now use `RawValue` everywhere.
The `inner_produce_protobuf_span` function needs to match on
`serde_json::Value`, but we can deserialize the raw values there. The
same applies for setting the sample rates in `backfill_data`.

#skip-changelog