feat(logs): Track accepted bytes per individual log item (#4886)
Calculates the ingested bytes when deserializing individual log items,
materializes it on the log and uses it for all outcomes which are to be
emitted.

There are quite a bit of follow ups related to this:
- #4881
- #4884
- #4885 
- #4887

See the tracking issue #4824 for details on the chosen algorithm to
calculate the ingested bytes.

Additionally, this PR does implement proper counting for **any** data
structure sent as an attribute value, this includes mixed arrays and
objects. Relay will drop these attributes later on as they are not yet
supported. This is generally the theme, Relay counts before making any
modifications. That could be because of size limits, PII,
inconsistencies etc.

This is inline with the approach we wanted to take with counting
received bytes. In practice this should not be a large issue, as people
are inclined to only send valid data, as the invalid data would not make
it to the product and SDKs are safeguarding against invalid usage as
well.

Closes: RELAY-111 
Closes: #4824