feat(relay): add mechanism to only allow events from trusted relays (#4772)
This PR adds a signature to requests between relays which is used to
verify if the request comes from a trusted relay.

It leverages the signature mechanism used by `relay-auth` which already
includes a timestamp which can be checked for a `max_age`. The signature
itself does not contain any data since we are not interested in data
integrity and we just want to see if we can successfully verify the
signature by any of the public keys that are added by the user.

Envelopes coming from internal relays are not checked because they are
already in our infrastructure and are considered trusted.

The `max_age` for envelopes from external relays is set to 5 Minutes
currently.

ref RELAY-17