fix(context): Overwrite `{}` and `[]` context type
Currently we only set the `"type"` field of a context if it's missing
entirely. However, we have seen cases in production where the type is
set to `{}`, which causes problems down the line. Therefore we now treat
a `"type"` of `{}` or `[]` the same as if it were missing. This seems
like a reasonable extension in which we definitely don't overwrite any
valid `"type"`.

A simpler alternative would be to use `value.is_empty()`, but then we
would also overwrite `""` types and it's not completely clear to me if
we want that. In any case that can be done as a follow-up.