feat(otlp): Serialize OTEL span array attributes to JSON (#4930)
OTLP supports array attributes, and there are many common semantic
attributes that are expected to be arrays. Right now, we discard this
information because we don't have system-level handling for arrays at
all, and a proper solution is very hard.

For now, let's hack around this, and serialize array types to a JSON
string, which is a common recommended approach by OTel when a protocol
cannot support every data type.

Convert array attributes _very early_ to strings, and then sends them
along.

---------

Co-authored-by: Sebastian Zivota <loewenheim@users.noreply.github.com>