ci(artifacts): No longer store debug/source artifacts in docker (#4942)
Simplifies the uploading of debug symbols and source files to gocd.

- No longer add them to the image itself, this should save quite a bit
of size of the image (~130-150 MiB)
- Use the previously uploaded artifacts to push them to gocd
- No longer upload libraries from the docker image, this seems to have
been largely not used, and is following Symbolicator, if it turns out
the symbols are useful, we can always add them back. I suspect most of
the libraries do not contain debug information anyways.

This is a preparation for using a distroless base image (#4871), which
will not contain (m)any system libraries either.