ref(quotas): Enforce span limits for transactions and vice versa (#4963)
Propagates and enforces span rate limits for transactions in the fast
path and vice versa.

This only implements the enforcement in the fast path and propagation,
it does *not* implement the enforcement side for consistent rate limits.
This is a different beast to implement and currently does not touch it,
this to be done in a follow up.

For motivation and justification:
https://develop.sentry.dev/ingestion/relay/transaction-span-ratelimits/

Refs: #4961