Rename legacy `ai.*` attributes to `gen_ai.*` names. (#4924)
This makes the names OTel compatible and also allows older SDKs that use
the legacy names to send spans that will appear in the AI agents
insights module.

Those are matching the deprecated `ai.*` attributes described in the
Sentry conventions:

https://getsentry.github.io/sentry-conventions/generated/attributes/ai.html#deprecated-attributes

We have checked if those attributes are used by customers in alerts,
dashboards and saved queries. They are NOT used, so we can safely
normalize them.

Fixes https://github.com/getsentry/sentry-python/issues/4516