feat(session): Add `unhandled` session status type (#4939)
This is in service of this project:
https://www.notion.so/sentry/RFC-Session-Status-unhandled-2308b10e4b5d801d9805fece61533ac5#2308b10e4b5d80e9b312ecbea27a4af3

**TL/DR:**
We're adding a new enum value to split up the existing `crashed` release
health metric into two more specific metrics:
- Unhandled
- Crashed
For cases like Flutter: a Dart error will not crash the running process,
while a native error will. We want to be clearer about this when we
report errors to sentry users.

**This PR**
This PR will allow us to properly count `session` and `sessions`
envelopes that contain the new value. When a `session` envelope arrives
that has the new value: `status: unhandled` we want the `SessionLike`
metric to handle that the same way we handle `status: abnormal` or
`status: crashed`. We'll subtract 1 from the `errors` count and return
`1` from `unhandled_count()`.

Similarly, for `sessions` envelopes we need to implement the
`unhandled_count()` function.

Finally, we have to implement the if-statements that will get the new
count from the session, `session.unhandled_count()` and push the value
into the list.


Fixes REPLAY-549

---------

Co-authored-by: Colton Allen <colton.allen@sentry.io>
Co-authored-by: Colton Allen <cmanallen90@gmail.com>
Co-authored-by: David Herberth <david.herberth@sentry.io>