fix(timestamps): Deal with sub-microsecond inaccuracies on parse (#5002)
This also deals with floating point sub-microsecond inaccuracies on
parsing, instead of only on serialization.

Logs access the nanoseconds of the extracted timestamp and persist these
into the payload, this leads to subtle inaccuracies because it never
goes through the floating point serialization code which would deal with
these inaccuracies.

As it turns out this also applies to calculated span durations which
were off.

Instead account for these inaccuracies when parsing from a float, but
also when serializing, as the timestamp may have been constructed from
something with 'perfect' accuracy.

Also fixes this:
<img width="668" height="100" alt="image"
src="https://github.com/user-attachments/assets/5105d3ca-b643-4305-9bb6-620413fff181"
/>