ref(ourlogs): Graduate metadata feature flag (#5006)
This has been since fully rolled out. We can graduate the flag by
deleting it, because this was only relevant for processing Relays and we
do not need to propagate it.

Removes a now redundant log, as well as removing the `debug_assert` for
capacity as it depends too much on the amount of metadata to add.



#skip-changelog