ref(ourlogs): Split processing into normalization and scrubbing (#5003)
Another preparation for inbound filters. We want to run inbound filters
before PII scrubbing, this does save some CPU cycles, but more
importantly this currently matches the behavior of errors.

This does lead to slightly inconsistent filters depending on in which
state they are executed (e.g. a filter that matches on a replaced value
will sometimes filter, sometimes not). But if we want to change this, we
should think about this more and change it for all item types.

Potentially in the future we may want to make a 'sub pipeline' which
processes individual logs one by one: `normalize -> filter -> scrub`. I
opted for this for now, as it matches more closely what we do for spans
and transactions. This is a simple refactor if need be.

#skip-changelog