ref(auth): add support for prehashed signature creation (#5012)
This PR introduces support for pre-hashed signatures.

Currently we create signatures by cloning the passed data, which in some
cases is the entire request body.
With pre-hashed signatures we can create the hash that is used in the
signature process ourselves by incrementally digesting it.
Unfortunately, signatures creates by the two mentioned mechanisms are
not compatible with each other.

Support for the new type is created by extending the signature header
with informations about how the signature was created. To remain
backwards compatible, it defaults to the old way of verifying signatures
in case this information is missing.

To allow for a smooth rollout, only the verification is introduced in
this PR and changing the creation will be done in a separate PR once
this is sufficiently rolled out.

ref INGEST-430