ref(rate-limits): Add convenience trait to rate limit by quantities (#5030)
Slightly changes rate limiting to logic to abort on the first breached
rate limit instead of all. I could not come up with a reason why doing
this would be disadvantageous and it will always be easier to later (if
we deem it necessary) to evaluate all. Doing the reverse requires more
considerations, since this code is completely new and unused, I think we
should start with this option.

The introduced type is also quite a generic type salad, but apart from
the error handling it isn't that bad and it reduces boilerplate, also we
end up with a single implementation used for multiple types preventing
bugs in the future.