feat(relay): Healthcheck command for self hosted and devservices (#5044)
We're (on self-hosted) trying to add healthcheck on Relay since David
changed the base image to distroless, hence the only thing on the Docker
image is just the relay binary.

I don't know how to pass in the config file while making it optional. If
this is specific for self-hosted, I don't think reading from the config
file is mandatory since we will be executing this on Docker Compose's
healthcheck mechanism.

---------

Co-authored-by: David Herberth <github@dav1d.de>
Co-authored-by: David Herberth <david.herberth@sentry.io>