feat(spanv2): Implements metric extraction and improves dynamic sampling (#5058)
Implements span metric extraction for SpanV2 items.

This is the second part to dynamic sampling which was missing, items
which are dropped must still have metrics extracted. Now, accepted
counts (calculated from the metrics) and the dynamic sampling
rebalancing metric (count per root) are correctly extracted.

The PR includes a minor refactor to `Managed` and `Output` to be able to
split off metrics and still account correctly for all outcomes and also
have a way to wrap metrics in a `Managed`.

There is also a slight change how metrics are sent to the sampling
project, this now includes an additional safeguard to not send the
metrics to the wrong project (to be verified if this is only possible
with the current implementation or also was happening before, there is
now a test for spanv2 which asserts that this does not happen).

A slight modification to how `SourceQuantities` are calculated, this was
a bug before, transactions and spans which are extracted from an indexed
payload should not have counted towards the outcome count, luckily this
was not an issue before because rate limits would've never been enforced
on these metrics in the first place.

---------

Co-authored-by: Sebastian Zivota <loewenheim@users.noreply.github.com>