feat: Add relay-conventions crate (#5049)
This adds a new workspace crate to parse the attribute definitions in
the `sentry-conventions` repository at compile time and provide them to
the rest of Relay. For this purpose it contains `sentry-conventions` as
a submodule.

Attribute definitions are parsed, converted to a more usable format, and
stored in a static map. The intention is to provide functions to query
different information about an attribute, and `write_behavior` is the
first such function.

All of the naming, internal structure, which information about
attributes should be included &c., is still massively subject to change
and I welcome any and all bikeshedding.

Note: this PR is currently blocked on
https://github.com/getsentry/sentry-conventions/issues/100.

ref: #5048. ref: RELAY-149.

---------

Co-authored-by: David Herberth <david.herberth@sentry.io>