feat(kafka): Always put transactions on transactions topic (#5081)
Minor improvement to the separation of error and transaction infra:

Error events with attachments are produced on the `attachments` topic to
[guarantee processing
order](https://github.com/getsentry/sentry/blob/64f5bb1f450fff0a31ea381fcaeee12965f50553/src/sentry/ingest/consumer/factory.py#L112-L115)
(e.g. for minidumps). As a side effect, transactions with attachments
were also produced on the `attachments` topic. Transaction attachments
are currently a rarely used feature which does not require ordering
guarantees, so we might as well produce the transactions themselves to
the `transactions` topic to separate them from errors.

Caveat here is that the _attachments_ of the transactions are still
produced onto the `attachments` topic, and are thus not separated from
error attachments.

Resolves INGEST-98, https://github.com/getsentry/team-ingest/issues/607