feat(PII): Add dynamic PII derivation (#5107)
This adds the ability to "dynamically" define PII to `metastructure`. It
works like this: if `compute_pii` is a function `fn(&ProcessingState) ->
Pii`, you can annotate a container or field with `#[metastructure(pii =
"compute_pii")]` (in addition to the existing values of `"true"`,
`"false"`, and `"maybe"`) which will cause the PII value of the
field/container to depend on the current `ProcessingState`. See the new
`test_dynamic_pii` test for a simple example of this feature in action.

The intended use of this is to allow the PII status of
`AttributeValue::value` to be fetched from `relay-conventions` based on
the attribute's name.

ref: #5096. ref: RELAY-152.