ref(pii): Change attr default/advanced rule logic (#5112)
This changes how the `PiiProcessor` detects whether attributes should be
scrubbed as objects (used for advanced rules) or identified with their
`value` field (used for default rules). Previously, this was decided
based on the presence of a segment of type `$log` in the processing
state. Now there is an explicit mode setting. This doesn't change the
behavior of the processor at all, it just makes the logic explicit and
applicable to other objects than logs.

ref: #5096. ref: RELAY-152.