feat(ai): Use fuzzy pattern matching as a fallback for matching model name (#5106)
Since snapshots of models change quite often, but the prices remain the
same, in case when we don't have exact match for the model name, we
might try to match it via glob. When generating global config with
prices, we will prepare the data to also include names with `*` pattern,
so that it can be used as a fallback.

E.g. `gpt-4-20250131` will also be saved as `gpt-4-*` meaning that all
the other snapshots of the models will be covered, even if there is not
exact match in the name in our pricing data.