fix(spanv2): Map `error` status to `internal_error` (#5127)
Currently, OTLP's `ERROR` status (`2`) and SpanV2's `error` status are
mapped to the SpanV1 status `unknown`. Since an `error` status means you
definitely had an error, the `unknown` status ("We do not know whether
the transaction failed or succeeded") does not seem appropriate.

The closest candidate for a generic error in SpanV1's status enum is
`internal_error`. Update the V2 to V1 mapping to use that instead.