feat(ai): Add ai_operation_type_map to global config (#5125)
### Description

In relay, we want to infer the `gen_ai.operation.type` attribute (it's a
newly added attribute) from `gen_ai.operation.name`.

This will be used in UI to easier query and plot data based on operation
type. Currently we have a [long
list](https://github.com/getsentry/sentry/blob/fd9d6f63b991bd4170defb3da1e0f18e7283cfd6/static/app/views/insights/agents/utils/query.tsx#L3)
of attributes that we query for one operation type, but as that list
grows, it is becoming harder and harder to pass all the parameters as
query param in a request to EAP.

### Implementation plan

PRs will be separated to make it easier to review. The mapping will be
defined in sentry, and it will be propagated to relays as a part of
global config to make it easier for us to change and extend the
mappings.

(in relay):
- [x] add `ai_operation_type_map` to global config (this PR)
- [ ] implement inference of the operation type from name and add tests

(in sentry):
- [ ] during global config generation, generate `ai_operation_type_map`
from defined list of attributes


Part of [TET-1092: Introduce `gen_ai.operation.type`
attribute](https://linear.app/getsentry/issue/TET-1092/introduce-gen-aioperationtype-attribute)